refactor(ui): extract key handlers to separate module
Move all keyboard input handling logic from mod.rs to key_handlers.rs
to improve code organization and maintainability.

Changes:
- New file: src/app/key_handlers.rs (662 lines)
- Updated: src/app/mod.rs (-634 lines, +37 lines)
- Made 28 App fields pub(crate) for handler access
- Made 26 helper methods pub(crate) for handler/test access
- Fixed unsafe array indexing in close_active_session()
- Made TextInput and ToolListState pub(crate) for consistency

Result:
- mod.rs reduced from 3,493 → 2,859 lines (19.2% reduction)
- Key handling logic now isolated in dedicated module
- Zero behavioral changes
- All 329 tests pass
- No clippy warnings
- All 6 architecture tests pass

This refactoring establishes a cleaner separation of concerns and makes
the codebase easier to navigate and maintain. The handlers module is now
a cohesive unit that can be independently reasoned about and tested.
